Fix: Created custom hook for lockout to be usable across the application.
